Added a 'non-urgent route caching method', that is designed to be performed over multiple ticks, for when the routes are not urgently required. It is always limited to performing a maximum of 0.1ms of calculation per world per tick.

+++ src/main/java/thermaldynamics/multiblock/Route.java
@@ -62,7 +62,7 @@ public class Route implements Comparable<Route> {
         return new Route(this);
     }
 
-    public int getNextDirection() {
+    public byte getNextDirection() {
 
         return pathDirections.get(pathPos++);
     }
